gen-ai-dev/projects/text_to_sql/requirements.txt
pandas
numpy
streamlit==1.36.0
openai==1.35.13
langchain==0.2.7
langchain-openai==0.1.16
faiss-cpu==1.8.0
langchain_community==0.2.7
pymysql==1.1.1
sqlalchemy
pybigquery
google-cloud-bigquery
sqlalchemy.orm
